refactor: move homepage component to it's own folder
