+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { Routes, Route } from 'react-router-dom';
-import Home from './pages/Home';
+import Home from './pages/Home/HomePage';
 import ProductDetails from './pages/ProductDetails';
 import Products from './pages/Products';
 import Header from './components/Header/Header';
